File reorganization, todo neighbor logic
